ctmap: add --time-diff option to show relative time
This patch adds an --time-diff option for to show the relative
expiration time (compared to now) when dumping CT entries with `cilium
bpf ct list`.

A first iteration of the patch changed the default output without adding
an option. For this version, the IPv4 fragmets test had to be changed because it
used the existing format. The current test is more robust than the
previous one, so I ended up keeping the change even though it is not
needed.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>